added probatron external tool

+++ src/main/java/eu/scape_project/hadoop/ConversionMapper.java
@@ -35,6 +35,7 @@ public class ConversionMapper extends MapReduceBase implements Mapper<LongWritab
         outdir = job.get("outdir");
         try {
             probatronJAR = new LocalFile(tempdir + "/probatron.jar");
+            if(new File(probatronJAR.getAbsolutePath()).exists()) { return; }
             InputStream in = ConversionRunner.class.getResourceAsStream("/external-tools/probatron.jar");
             OutputStream out = new FileOutputStream(new File(probatronJAR.getAbsolutePath()));
             byte[] buf = new byte[1024];
@@ -43,7 +44,8 @@ public class ConversionMapper extends MapReduceBase implements Mapper<LongWritab
             in.close();
             out.close();
 
-            probatronSchema = new LocalFile(tempdir + "kbMaster.sch");
+            probatronSchema = new LocalFile(tempdir + "/kbMaster.sch");
+            if(new File(probatronSchema.getAbsolutePath()).exists()) { return; }
             in = ConversionRunner.class.getResourceAsStream("/kbMaster.sch");
             out = new FileOutputStream(new File(probatronSchema.getAbsolutePath()));
             buf = new byte[1024];
